Move EditHomePage to containers
