+++ front/app/containers/Admin/pagesAndMenu/containers/EditHomepage/index.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from 'react';
 
 // components
-import SectionToggle from '../SectionToggle';
+import SectionToggle from '../../SectionToggle';
 import { Box, Title } from '@citizenlab/cl2-component-library';
 import Warning from 'components/UI/Warning';
 import AdminViewButton from './AdminViewButton';
-import SectionFormWrapper from '../components/SectionFormWrapper';
-import { pagesAndMenuBreadcrumb, homeBreadcrumb } from '../breadcrumbs';
+import SectionFormWrapper from '../../components/SectionFormWrapper';
+import { pagesAndMenuBreadcrumb, homeBreadcrumb } from '../../breadcrumbs';
 
 // i18n
 import messages from './messages';
